rss_x/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-cors>=4.0.0
sqlalchemy>=2.0
pydantic>=2.0
feedparser>=6.0
requests>=2.31
PyYAML>=6.0
python-dotenv>=1.0
pyjwt>=2.0
bcrypt>=4.0
openai>=1.0
curl_cffi>=0.7.0
pytest>=8.0
pytest-cov>=4.0
